+++ Etyllica/src/br/com/etyllica/gui/icon/DownArrow.java
@@ -2,17 +2,17 @@ package br.com.etyllica.gui.icon;
 
 public class DownArrow extends PolygonalIcon{
 
-	public DownArrow(int x, int y, int size){
+	public DownArrow(float x, float y, float size){
 		super(x, y, size);
 				
 	}	
 	
-	protected void initPolygon(int x, int y){
+	protected void initPolygon(float x, float y){
 		polygon.reset();
 		
-		polygon.addPoint(x, y);
-		polygon.addPoint(x+size, y);
-		polygon.addPoint(x+size/2, y+size);
+		polygon.addPoint((int)x, (int)y);
+		polygon.addPoint((int)(x+size), (int)y);
+		polygon.addPoint((int)(x+size/2), (int)(y+size));
 	}
 
 }
