+++ Etyllica/src/br/com/etyllica/gui/label/Icon.java
@@ -4,7 +4,7 @@ import br.com.etyllica.gui.Label;
 
 public abstract class Icon extends Label{
 
-	public Icon(int x, int y){
+	public Icon(float x, float y){
 		super(x,y);
 	}
