+++ Etyllica/src/examples/jgl/application/BezSurf.java
@@ -79,7 +79,7 @@ public class BezSurf extends GL{
 	}
 
 
-	private void myReshape (int w, int h) {
+	private void myReshape (float w, float h) {
 		glViewport (0, 0, w, h);
 		glMatrixMode (GL_PROJECTION);
 		glLoadIdentity ();
